+++ src/main/java/duke/Duke.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -5,6 +7,8 @@ import java.util.Collections;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
+import duke.task.*;
+import duke.exception.*;
 
 enum TaskInfo {
     TASKNAME,
