+++ src/main/java/duke/task/Events.java
@@ -1,15 +1,16 @@
+package duke.task;
 
-public class Deadlines extends Task {
+public class Events extends Task {
     private final String dateTime;
 
-    private Deadlines(int taskId, String taskName, 
+    private Events(int taskId, String taskName, 
             String dateTime, boolean isDone) {
 
         super(taskId, taskName, isDone);
         this.dateTime = dateTime;
     }
 
-    public Deadlines(int taskId, String taskName, String dateTime) {
+    public Events(int taskId, String taskName, String dateTime) {
         super(taskId, taskName);
         this.dateTime = dateTime;
     }
@@ -20,23 +21,23 @@ public class Deadlines extends Task {
 
     @Override
     public String taskWithSymbol() {
-        return "[D]" 
+        return "[E]" 
                 + ((this.isDone) ? "[✓]" : "[✗]")
-                + " " 
+                + "   " 
                 + this.taskName;
     }
 
     @Override
-    public Deadlines makeDone() {
-        return new Deadlines(this.taskId, this.taskName, 
+    public Events makeDone() {
+        return new Events(this.taskId, this.taskName, 
                 this.dateTime, true);
     }
 
     @Override
     public String toString() {
         return this.taskId 
-                + "." 
+                + "."  
                 + this.taskWithSymbol()
-                + " (by: " + this.dateTime + ")";
+                + " (at: " + this.dateTime + ")";
     }
 }
