+++ src/main/java/duke/task/ToDos.java
@@ -1,3 +1,4 @@
+package duke.task;
 
 public class ToDos extends Task {
 
@@ -13,7 +14,7 @@ public class ToDos extends Task {
     public String taskWithSymbol() {
         return "[T]" 
                 + ((this.isDone) ? "[✓]" : "[✗]")
-                + " " 
+                + "   " 
                 + this.taskName;
     }
